+++ src/public/vgui/IInput.h
@@ -21,7 +21,7 @@ namespace vgui
 {
 
 class Cursor;
-typedef uint32 HCursor;
+typedef unsigned long HCursor;
 
 #define VGUI_GCS_COMPREADSTR                 0x0001
 #define VGUI_GCS_COMPREADATTR                0x0002
